Updating to meet PR comments.
