+++ src/pdf_parser/pdf_utils/disambiguator/tests/test_utils.py
@@ -1,12 +1,10 @@
 import pytest
 
-from src.pdf_parser.pdf_utils.disambiguator.unexplained import (
-    calculate_unexplained_fractions,
-)
 from src.pdf_parser.pdf_utils.disambiguator.utils import (
     is_in,
     split_layout,
     combine_layouts,
+    calculate_unexplained_fractions,
 )
 
 
@@ -52,3 +50,15 @@ def test_combine(test_layout_restrictive, test_layout_random):
     assert len(permissive_layout) + len(
         [i for i in unexplained_fractions if i > 0.5]
     ) == len(combined_layout)
+
+
+@pytest.mark.parametrize("test_layout_random", [100], indirect=["test_layout_random"])
+def test_calculate_unexplained_fractions(test_layout_random, test_layout_restrictive):
+    """Tests that the unexplained fraction is calculated correctly."""
+    permissive_layout, unexplained_fractions = calculate_unexplained_fractions(
+        restrictive_layout=test_layout_restrictive, permissive_layout=test_layout_random
+    )
+
+    assert all([0 <= f <= 1 for f in unexplained_fractions])
+
+    assert len(unexplained_fractions) == len(permissive_layout)
